Fix 2 failing specs by adding message value to test struct
Two specs fail as there is no `message` member in the Exception returned
by a mocked function call.

```
  1) Marloss::Store.create_lock should fail creating the lock and raise error
     Failure/Error: expect { store.create_lock(name) }.to raise_error(Marloss::LockNotObtainedError)

       expected Marloss::LockNotObtainedError, got #<NameError: no member 'message' in struct> with backtrace:
         # ./lib/marloss/store.rb:112:in `rescue in create_lock'
         # ./lib/marloss/store.rb:89:in `create_lock'
         # ./spec/lib/marloss/store_spec.rb:115:in `block (4 levels) in <top (required)>'
         # ./spec/lib/marloss/store_spec.rb:115:in `block (3 levels) in <top (required)>'
     # ./spec/lib/marloss/store_spec.rb:115:in `block (3 levels) in <top (required)>
```

Add the `message` value to the mocked exception.

Clearly this must work without this message being set under some
circumstances or it would not be merged into the repo. I am using the
versions of the gems from the Gemfile, and a freshly compiled Ruby 2.4.3 .