Only release lock if we obtained it (#17)
`with_marloss_locker` has an ensure section that makes sure that we
release the lock even if an error occurs.

It does not guarantee that we had actually obtained the lock. Thus it
will release a lock that another process has obtained. This will happen
in the case of an attempt to obtain the lock failing due to max retries,
and raising Marloss::LockNotObtainedError.

Ensure that we have the lock before we release it.